Renamed ValidateCurrentSession to reflect quickstart
